Refactor animation duration types to double
Updated `AnimationHelpers` and related classes to use `double` for duration parameters, enhancing precision in animations. Removed `AnimationNavigationControllerDelegate` and simplified `TransitionAnimator` by eliminating navigation awareness logic. Adjusted `ShellTransSectionRenderer` and `ShellTrans` for consistency with new duration types. Added a new solution file for project configuration.

+++ PJ.NavigationTransitions.Maui/AnimationHelpers.android.cs
@@ -25,13 +25,13 @@ static partial class AnimationHelpers
 		_ => Resource.Animation.none,
 	};
 
-	public static AnimationInfo ToPlatform(this TransitionType transition, int duration)
+	public static AnimationInfo ToPlatform(this TransitionType transition, double duration)
 	{
 		var animation = transition.ToPlatform();
 		var context = Platform.AppContext ?? throw new NullReferenceException();
 		var loadedAnimation = Android.Views.Animations.AnimationUtils.LoadAnimation(context, animation);
 		Debug.Assert(loadedAnimation is not null);
-		loadedAnimation.Duration = duration;
+		loadedAnimation.Duration = (long)duration;
 		return new(animation, loadedAnimation);
 	}
 }
